backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic==2.9.0
pydantic-settings==2.5.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
chromadb==0.5.0
sentence-transformers==3.0.0
python-multipart==0.0.9
httpx==0.27.0
